cfgvendor: drop the redundant size_t cast in the ring_name bound
min_t() already assigns both operands to the given type, so casting
nla_len() adds nothing.

+++ os_dep/linux/rtw_cfgvendor.c
@@ -1372,7 +1372,7 @@ static int rtw_cfgvendor_logger_start_logging(struct wiphy *wiphy,
 #if (LINUX_VERSION_CODE >= KERNEL_VERSION(7, 2, 0))
 				strscpy_pad(ring_name, nla_data(iter),
 					min_t(size_t, sizeof(ring_name),
-					      (size_t)nla_len(iter) + 1));
+					      nla_len(iter) + 1));
 #else
 				strncpy(ring_name, nla_data(iter),
 					MIN(sizeof(ring_name) -1, nla_len(iter)));
@@ -1506,7 +1506,7 @@ static int rtw_cfgvendor_logger_get_ring_data(struct wiphy *wiphy,
 #if (LINUX_VERSION_CODE >= KERNEL_VERSION(7, 2, 0))
 				strscpy_pad(ring_name, nla_data(iter),
 					min_t(size_t, sizeof(ring_name),
-					      (size_t)nla_len(iter) + 1));
+					      nla_len(iter) + 1));
 #else
 				strncpy(ring_name, nla_data(iter),
 					MIN(sizeof(ring_name) -1, nla_len(iter)));
